Starting code for a socket-based timing provider implementation
This commit contains the skeleton of a WebSocket server, and of a
client SocketTimingProvider class that would connect to it.

The commit also drops the "fetch" command on the TimingProvider class
since it seems useless in the end, implements "readyState" on the TimingObject
class and improves the dispatching of "change" events using setters.

The socket code is not yet up and running.