The "change" event is now only fired when vector has changed for real
Setting a new "vector" on a timing provider object will now only fire a
"change" event provided the vector changed for real, i.e. when the
position, velocity or acceleration evaluated at some timestamp is
different from the current vector.

Similarily, associating a timing object with a timing provider will not
trigger a "change" event if the new vector is the same as the previous
one.