First crude clock synchronization mechanism added
The current code merely computes the skew every second, without considering
past values and round trip times. This is rather basic and not very good from
a synchronization perspective!

The local vector is updated each time the skew is updated.

Note that change messages that lie in the future according to the client are
not queued for later propagation but rather executed immediately. To be fixed
in a later commit.