Moved sync logic to a separate SyncClock module
Separating the logic that deals with the motion vector and the logic
that deals with the clock synchronization seems logic in the code. It
also makes it possible to use some external synchronized clock (in other
words one that is not bound to the same socket connection for instance).

Ready states "opening", "open", "closing", and "closed" have also been
implemented correctly throughout the code.