Added logic to delay execution of changes received from the server
The server may send updates that are to be executed in the future,
because it is able to predict a change, or because it requires all
clients to delay changes by a certain amount of time to improve
synchronization of updates across clients, or because the skew
estimation indicates that the change received should actually happen
later on.

The code now queues such updates and schedules them for later execution.

The socket server may now also be started with a delta parameter that
sets the delay that all clients must apply to incoming change events.
For instance, to start the server with a 2 seconds delay:

 node src/server.js 2000