Client code now runs in browsers. First browser examples defined
The code should now run correctly in browsers, provided one uses Require.js.
See the first couple of basic examples in "examples/browser".

Please note that the console still reports an initial error. That is due to
the fact that the Websockets library is not compatible with AMD and thus
cannot be "required" from within a Web browser. This error is caught and
the client side code simply uses the native WebSocket implementation when
it is available.