Improved socket clock synchronization using an initialization period
The socket clock now only switches to the "open" ready state at the end of
an initialization period during which it sends a series of 10 synchronization
requests to the socket server, which it then uses to determine the initial
skew but also the minimum roundtrip duration and a realistic roundtrip
threshold.

The clock then switches to the "open" mode when it tries to synchronize
with the online server once in a while (every 10 seconds currently).
To synchronize with the online server, the clock sends up to 10 "sync"
requests to the online server and only considers the responses to be
valid when the response is received within the computed threshold.

This behavior rejects messages that spent too much time lost in limbos
(in the network or waiting to be processed on the client or server)
and that would result in a poor estimate of the local skew.

If synchronization fails, the clock simply considers that the current
skew is still valid.

Since the skew estimate should be considered valid until next time,
the socket timing provider does not attempt to change the "timestamp"
of the current state vector anymore. It still does that for pending
state vectors.