Added basic static file serving capabilities to HTTP/Websocket server
The HTTP server exposed by the Websocket server now serves static files
in the "examples" folder. Also, the browser examples that use a WebSocket
now use the underlying host name with which the HTML page is loaded to
compute the WebSocket URL.

This makes it easy to expose the HTTP URL on other addresses than localhost,
needed to load a given example from multiple computers at once.