Socket clock now correctly calls addEventListener when possible
The socket clock needs to listen to events emitted by the socket connection
but the timing provider object will typically also listen to the same events
as well. The WebSocket library used for the Node.js environment does not
support "addEventListener", so the code worked around.

The code now uses "addEventListener" when possible, in Web browser
environments in particular.