Fixed bug in socket-based clock that could lead to repeated sync attempts
There existed a possibility where the timeout interval set to detect messages
that take too long was triggered before the message from the clock server was
received but where the message from the clock server was still processed as
valid. When that happened, sync attempts started to behave erratically,
leading to lots of sync requests being sent per second.